feat: update the set-audio-input script
Get the node's name from the device.description if a node.nick is not available.
Do not filter by available nodes, since this is not reported for filter nodes (e.g. bluetooth mic.) nor by the analog jack mic.
Get the nodes' names by filtering the whole pactl output with the found object.ids instead of finding the object.ids and the names 'at the same time'.
Several variable renames.