add three fzf related shell scripts
fzf-popup-window opens a floating kitty terminal window, allowing a single instance to be opened.
fzf-xdg-open shows a list of files in the home dir. and allows one of them to be opened with xdg-open.
fzf-xdg-mime-helper returns the default app for a given file; this is used within the previous script.