notify-ext-scr-brightness script: allow faster notifications
This update makes use of a file to store the brightness levels locally.

If the file doesn't exist, the file's contents don't have the right
format, or if the file's data is not recent enough, a new query
of the current monitor brightness will be done and stored.

The idea is that, for consecutive calls of the script, the brightness
is not queried from the monitor every time, as this is quite slow for
ddcutil, but rather calculated from the local data.