less.vim: change the used mappings
This change is mainly motivated by the fact that 'd' and 'g' have
a bunch of extra native mappings, so that if one wants to use them on
their own in this 'less mode', a noticeable delay is introduced.

So 'g' was changed to the usual 'gg' and 'd' is not used anymore.
To keep it consistent, 'b', 'f' and 'u' are also removed.

Additionally, a bunch of editing mappings that were unmapped.

The help text was updated accordingly.