fzf-popup-window: update the logic of the script
Instead of passing a script file as an argument and assuming that the
script will have some fzf functionality in it, this new version works
more like the normal fzf, in that the options to be shown have to be
piped to the script, and the selection will simply be piped out.

Other scripts that rely on this one will break until updated.